+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 6.19.5
 
+* fixed reading imported statistics with platform filter
 * updated dependencies
 
 ## 6.19.4